Renamed salt_observation_properties.py file to observation_properties.py
