+++ src/ssda/observation_properties.py
